IT-5050 | rename SELECT_LIST_SIZES to SELECT_LIST_ITEM_SIZES
