+++ lib/js/components/SelectList/SelectListItem/SelectListItem.stories.ts
@@ -2,7 +2,7 @@ import SelectListItem from './SelectListItem.vue';
 import { ICONS } from '../../Icons/Icon';
 
 import { Args, ArgTypes, Meta, StoryFn } from '@storybook/vue';
-import { SELECT_LIST_ITEM_STATES, SELECT_LIST_SIZES } from '../SelectList.consts';
+import { SELECT_LIST_ITEM_STATES, SELECT_LIST_ITEM_SIZES } from './SelectListItem.consts';
 
 export default {
 	title: 'Components/SelectList/SelectListItem',
@@ -31,7 +31,7 @@ const StoryTemplate: StoryFn<typeof SelectListItem> = (argTypes) => ({
 export const Interactive = StoryTemplate.bind({});
 
 const args = {
-	size: SELECT_LIST_SIZES.SMALL,
+	size: SELECT_LIST_ITEM_SIZES.SMALL,
 	iconLeft: null,
 	label: 'Label',
 	state: SELECT_LIST_ITEM_STATES.DEFAULT,
@@ -43,7 +43,7 @@ const argTypes = {
 		control: { type: 'select', options: [null, ...Object.keys(ICONS)] },
 	},
 	size: {
-		control: { type: 'select', options: Object.values(SELECT_LIST_SIZES) },
+		control: { type: 'select', options: Object.values(SELECT_LIST_ITEM_SIZES) },
 	},
 	state: {
 		control: { type: 'select', options: Object.values(SELECT_LIST_ITEM_STATES) },
