--- lib/js/components/SelectList/index.ts
@@ -1,4 +1,0 @@
-import SelectList from './SelectList.vue';
-
-export * from './SelectList.consts';
-export default SelectList;
