+++ lib/js/index.ts
@@ -45,11 +45,11 @@ export { default as Pill } from './components/Pill';
 export * from './components/Pill/Pill.consts';
 export { default as CounterToggle } from './components/Toggles/CounterToggle';
 export * from './components/Toggles/CounterToggle/CounterToggle.consts';
-export { default as SelectList } from './components/SelectList';
+export { default as SelectList } from './components/SelectList/SelectList.vue';
 export { default as SelectListItem } from './components/SelectList/SelectListItem/SelectListItem.vue';
 export { default as SelectListItemDivider } from './components/SelectList/SelectListItemDivider/SelectListItemDivider.vue';
 export { default as SelectListItemToggle } from './components/SelectList/SelectListItemToggle/SelectListItemToggle.vue';
-export * from './components/SelectList/SelectList.consts';
+export * from './components/SelectList/SelectListItem/SelectListItem.consts';
 export { default as PopOver } from './components/PopOver';
 export * from './components/PopOver/PopOver.consts';
 export { default as Dropdown } from './components/Dropdown';
